requirements.txt
websockets
numpy
ipython
msgpack==1.0.0
PyTurboJPEG
